@@ -182,6 +182,8 @@ class KernelPCovRInfrastructureTest(KernelPCovRBaseTest):
         self.assertTrue(hasattr(kpcovr, "centerer_"))
         _ = kpcovr.predict(self.X)
         _ = kpcovr.transform(self.X)
+
+        print(self.Y.shape)
         _ = kpcovr.score(self.X, self.Y)
 
     def test_prefit_regressor(self):
@@ -255,7 +257,7 @@ class KernelPCovRInfrastructureTest(KernelPCovRBaseTest):
 
         # Dimension mismatch
         with self.assertRaises(ValueError) as cm:
-            kpcovr.fit(self.X, self.Y[:, 0])
+            kpcovr.fit(self.X, np.zeros(self.Y.shape + (2,)))
         self.assertTrue(
             str(cm.exception),
             "The regressor coefficients have a dimension incompatible "
